[VIVO-1245] Cleanup of DefaultAddMissingIndividualFormGenerator
